+++ apps/stripe/src/__tests__/mocks/app-config-repo.ts
@@ -2,7 +2,7 @@ import { ok } from "neverthrow";
 import { vi } from "vitest";
 
 import { mockedStripeConfig } from "@/__tests__/mocks/mock-stripe-config";
-import { AppConfigRepo } from "@/modules/app-config/app-config-repo";
+import { AppConfigRepo } from "@/modules/app-config/repositories/app-config-repo";
 
 export const mockedAppConfigRepo: AppConfigRepo = {
   getStripeConfig: async () => {
