+++ apps/stripe/src/__tests__/mocks/mock-stripe-config.ts
@@ -2,7 +2,7 @@ import { mockedConfigurationId } from "@/__tests__/mocks/constants";
 import { mockedStripePublishableKey } from "@/__tests__/mocks/mocked-stripe-publishable-key";
 import { mockedStripeRestrictedKey } from "@/__tests__/mocks/mocked-stripe-restricted-key";
 import { mockStripeWebhookSecret } from "@/__tests__/mocks/stripe-webhook-secret";
-import { StripeConfig } from "@/modules/app-config/stripe-config";
+import { StripeConfig } from "@/modules/app-config/domain/stripe-config";
 
 export const mockedStripeConfig = StripeConfig.create({
   id: mockedConfigurationId,
