+++ apps/stripe/src/modules/app-config/domain/app-root-config.ts
@@ -1,4 +1,4 @@
-import { StripeConfig } from "@/modules/app-config/stripe-config";
+import { StripeConfig } from "@/modules/app-config/domain/stripe-config";
 
 export class AppRootConfig {
   readonly chanelConfigMapping: Record<string, string>;
