+++ apps/stripe/src/modules/app-config/domain/stripe-config.ts
