+++ apps/stripe/src/modules/app-config/repositories/dynamodb/channel-config-mapping-db-model.ts
