+++ apps/stripe/src/modules/app-config/repositories/dynamodb/stripe-config-db-model.ts
