+++ apps/stripe/src/modules/app-config/trpc-handlers/get-stripe-config-trpc-handler.ts
@@ -1,7 +1,7 @@
 import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
-import { StripeFrontendConfig } from "@/modules/app-config/stripe-config";
+import { StripeFrontendConfig } from "@/modules/app-config/domain/stripe-config";
 import { createSaleorApiUrl } from "@/modules/saleor/saleor-api-url";
 import { protectedClientProcedure } from "@/modules/trpc/protected-client-procedure";
