+++ apps/stripe/src/modules/app-config/trpc-handlers/get-stripe-configs-channels-mapping-trpc-handler.test.ts
@@ -8,7 +8,7 @@ import { mockedStripeConfig } from "@/__tests__/mocks/mock-stripe-config";
 import { mockedSaleorApiUrl } from "@/__tests__/mocks/saleor-api-url";
 import { TEST_Procedure } from "@/__tests__/trpc-testing-procedure";
 import { BaseError } from "@/lib/errors";
-import { AppRootConfig } from "@/modules/app-config/app-root-config";
+import { AppRootConfig } from "@/modules/app-config/domain/app-root-config";
 import { GetStripeConfigsChannelsMappingTrpcHandler } from "@/modules/app-config/trpc-handlers/get-stripe-configs-channels-mapping-trpc-handler";
 import { router } from "@/modules/trpc/trpc-server";
