+++ apps/stripe/src/modules/app-config/trpc-handlers/get-stripe-configs-channels-mapping-trpc-handler.ts
@@ -3,7 +3,7 @@ import { TRPCError } from "@trpc/server";
 import {
   StripeFrontendConfig,
   StripeFrontendConfigSerializedFields,
-} from "@/modules/app-config/stripe-config";
+} from "@/modules/app-config/domain/stripe-config";
 import { createSaleorApiUrl } from "@/modules/saleor/saleor-api-url";
 import { protectedClientProcedure } from "@/modules/trpc/protected-client-procedure";
