+++ apps/stripe/src/modules/app-config/trpc-handlers/get-stripe-configs-list-trpc-handler.ts
@@ -1,6 +1,6 @@
 import { TRPCError } from "@trpc/server";
 
-import { StripeFrontendConfig } from "@/modules/app-config/stripe-config";
+import { StripeFrontendConfig } from "@/modules/app-config/domain/stripe-config";
 import { createSaleorApiUrl } from "@/modules/saleor/saleor-api-url";
 import { protectedClientProcedure } from "@/modules/trpc/protected-client-procedure";
